+++ tests/testapp/testapp/urls.py
@@ -1,27 +1,17 @@
 from __future__ import absolute_import, unicode_literals
 
-import wagtail
 from django.conf import settings
-from django.conf.urls import include, url
-
-if wagtail.VERSION >= (2, 0):
-    from wagtail.admin import urls as wagtailadmin_urls
-    from wagtail.core import urls as wagtail_urls
-    from wagtail.documents import urls as wagtaildocs_urls
-    from wagtail.images import urls as wagtailimages_urls
-else:
-    from wagtail.wagtailadmin import urls as wagtailadmin_urls
-    from wagtail.wagtailcore import urls as wagtail_urls
-    from wagtail.wagtaildocs import urls as wagtaildocs_urls
-    from wagtail.wagtailimages import urls as wagtailimages_urls
+from django.urls import include, re_path
+from wagtail import urls as wagtail_urls
+from wagtail.admin import urls as wagtailadmin_urls
+from wagtail.documents import urls as wagtaildocs_urls
+from wagtail.images import urls as wagtailimages_urls
 
 urlpatterns = [
-    url(r'^admin/', include(wagtailadmin_urls)),
-
-    url(r'^documents/', include(wagtaildocs_urls)),
-    url(r'^images/', include(wagtailimages_urls)),
-
-    url(r'', include(wagtail_urls)),
+    re_path(r"^admin/", include(wagtailadmin_urls)),
+    re_path(r"^documents/", include(wagtaildocs_urls)),
+    re_path(r"^images/", include(wagtailimages_urls)),
+    re_path(r"", include(wagtail_urls)),
 ]