+++ wagtailcaptcha/forms.py
@@ -1,26 +1,22 @@
 from __future__ import absolute_import, unicode_literals
 
-import wagtail
-from captcha.fields import ReCaptchaField
-
-if wagtail.VERSION >= (2, 0):
-    from wagtail.contrib.forms.forms import FormBuilder
-else:
-    from wagtail.wagtailforms.forms import FormBuilder
+from django_recaptcha.fields import ReCaptchaField
+from wagtail.contrib.forms.forms import FormBuilder
 
 
 class WagtailCaptchaFormBuilder(FormBuilder):
-    CAPTCHA_FIELD_NAME = 'wagtailcaptcha'
+    CAPTCHA_FIELD_NAME = "wagtailcaptcha"
 
     @property
     def formfields(self):
         # Add wagtailcaptcha to formfields property
         fields = super(WagtailCaptchaFormBuilder, self).formfields
-        fields[self.CAPTCHA_FIELD_NAME] = ReCaptchaField(label='')
+        fields[self.CAPTCHA_FIELD_NAME] = ReCaptchaField(label="")
 
         return fields
 
 
 def remove_captcha_field(form):
-    form.fields.pop(WagtailCaptchaFormBuilder.CAPTCHA_FIELD_NAME, None)
-    form.cleaned_data.pop(WagtailCaptchaFormBuilder.CAPTCHA_FIELD_NAME, None)
+    if form.is_valid():
+        form.fields.pop(WagtailCaptchaFormBuilder.CAPTCHA_FIELD_NAME, None)
+        form.cleaned_data.pop(WagtailCaptchaFormBuilder.CAPTCHA_FIELD_NAME, None)