+++ wagtailcaptcha/models.py
@@ -1,12 +1,8 @@
 from __future__ import absolute_import, unicode_literals
 
-import wagtail
-from .forms import WagtailCaptchaFormBuilder, remove_captcha_field
+from wagtail.contrib.forms.models import AbstractEmailForm, AbstractForm
 
-if wagtail.VERSION >= (2, 0):
-    from wagtail.contrib.forms.models import AbstractEmailForm, AbstractForm
-else:
-    from wagtail.wagtailforms.models import AbstractEmailForm, AbstractForm
+from .forms import WagtailCaptchaFormBuilder, remove_captcha_field
 
 
 class WagtailCaptchaEmailForm(AbstractEmailForm):